+++ CHANGELOG.md
@@ -254,6 +254,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Recover and use lost _preparePrice result in Article::_prepareModifiedPrice [PR-720](https://github.com/OXID-eSales/oxideshop_ce/pull/720)
 - Load amount price list same way on frontend and backend [#0006671](https://bugs.oxid-esales.com/view.php?id=6671) [PR-712](https://github.com/OXID-eSales/oxideshop_ce/pull/712)
+- Show product active check in admin panel when activation date has been set [#0006966](https://bugs.oxid-esales.com/view.php?id=6966)
 
 ## [6.3.5] - 2019-07-30